fix again

+++ cfr/visual.py
@@ -529,7 +529,7 @@ def plot_proxies(df, year=np.arange(2001), lon_col='lon', lat_col='lat', type_co
 
         proxy_count = {}
         for ptype in df_count.keys():
-            proxy_count[ptype] = df_count[ptype].sum(axis=1)
+            proxy_count[ptype] = df_count[ptype].sum(axis=1).astype(int)
 
         cumu_count = np.zeros_like(year, dtype=int)
         cumu_last = np.copy(cumu_count)
